start moving coroutines to a new module
